requirements.txt
alembic==1.13.1
blinker==1.7.0
cachelib==0.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
circuitbreaker==2.0.0
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Caching==2.1.0
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
logzero==1.7.0
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
packaging==24.0
proofdock-chaos-middleware-flask==0.1.3
proofdock-chaos-middleware-python==0.2.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.4
requests==2.31.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2